% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanSVG}
\alias{cleanSVG}
\title{Clean a SVG provided by RDKit}
\usage{
cleanSVG(svg, out.w, out.h)
}
\arguments{
\item{svg}{a svg}

\item{out.w}{output width}

\item{out.h}{output height}
}
\value{
a SVG
}
\description{
Cleaning conist on resizing
}

