% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inchi.R
\name{compareInchiKeys}
\alias{compareInchiKeys}
\title{Compare Inchikeys}
\usage{
compareInchiKeys(inchikey1, inchikey2, check.protonation = T)
}
\arguments{
\item{inchikey1}{A inchikey}

\item{inchikey2}{A inchikey}

\item{check.protonation}{A check protonation}
}
\value{
Bool
}
\description{
Compare Inchikeys
}

