% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{compute2D}
\alias{compute2D}
\title{Computes 2D coordinates for molecules}
\usage{
compute2D(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\description{
Computes 2D coordinates for molecules
}

