% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fragmentOnBRICSBonds}
\alias{fragmentOnBRICSBonds}
\title{split molecule to brics fragment}
\usage{
fragmentOnBRICSBonds(xp)
}
\arguments{
\item{xp}{a molecule}
}
\value{
a BRICS fragment
}
\description{
split molecule to brics fragment
}

