% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBRICSFragments}
\alias{getBRICSFragments}
\title{Get unique BRICS fragments for a molecule}
\usage{
getBRICSFragments(mol)
}
\arguments{
\item{mol}{a molecule}
}
\value{
Vector of BRICS fragments
}
\description{
Get unique BRICS fragments for a molecule
}

