% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getSMARTSFP}
\alias{getSMARTSFP}
\title{Get FP defined by SMARTS}
\usage{
getSMARTSFP(mols, smarts.mols)
}
\arguments{
\item{mols}{A list of molecules}

\item{smarts.mols}{A list of SMARTS molecules}
}
\value{
A bool matrix with one row per molecule and each column a SMARTS Fingerprint
}
\description{
Get FP defined by SMARTS
}

