% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inchi.R
\name{Inchi2InchiKey}
\alias{Inchi2InchiKey}
\title{Map Inchi to  Inchikeys}
\usage{
Inchi2InchiKey(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
A list of inchikeys
}
\description{
Map Inchi to  Inchikeys
}

