% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Inchi2Mol}
\alias{Inchi2Mol}
\title{Map Inchi to molecule}
\usage{
Inchi2Mol(x, sanitize = FALSE, removeHs = FALSE)
}
\arguments{
\item{x}{a  molecule}

\item{sanitize}{bool}

\item{removeHs}{bool}
}
\value{
a molecule
}
\description{
Map Inchi to molecule
}

