% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.molecule}
\alias{is.molecule}
\title{Check if object is a "nice" RDKit Molecule}
\usage{
is.molecule(mol)
}
\arguments{
\item{mol}{A molecule}
}
\description{
Check if object is a "nice" RDKit Molecule
}

