% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.nice.smile}
\alias{is.nice.smile}
\title{Check if object is a "nice" RDKit smile}
\usage{
is.nice.smile(smile)
}
\arguments{
\item{smile}{A smile}
}
\description{
Check if object is a "nice" RDKit smile
}

