% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kekulize}
\alias{kekulize}
\title{Kekulize molecule}
\usage{
kekulize(xp, markAtomsBonds = TRUE, maxBackTracks = 100L)
}
\arguments{
\item{xp}{a molecule}

\item{markAtomsBonds}{markAtomsBonds}

\item{maxBackTracks}{maxBackTracks}
}
\description{
Kekulize molecule
}

