% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inchi.R
\name{mol2InchiKey}
\alias{mol2InchiKey}
\title{Map mols to Inchikeys}
\usage{
mol2InchiKey(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
A list of inchikeys
}
\description{
Map mols to Inchikeys
}

