% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mol2maccs}
\alias{mol2maccs}
\title{map a molecule to a maccs fingerprints}
\usage{
mol2maccs(xp)
}
\arguments{
\item{xp}{a molecule}
}
\value{
a vector
}
\description{
map a molecule to a maccs fingerprints
}

