% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mol2morgan}
\alias{mol2morgan}
\title{map a molecule to a morgan fingerprints (atom environement)}
\usage{
mol2morgan(xp, radius = 2L, nBits = 2048L, useFeatures = FALSE)
}
\arguments{
\item{xp}{a molecule}

\item{radius}{radius of scaffold}

\item{nBits}{final length}

\item{useFeatures}{useFeatures}
}
\value{
a vector
}
\description{
map a molecule to a morgan fingerprints (atom environement)
}

