% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mol2murcko}
\alias{mol2murcko}
\title{Compute Murcko scaffold}
\usage{
mol2murcko(xp)
}
\arguments{
\item{xp}{a molecule}
}
\value{
the Murcko scaffold
}
\description{
Compute Murcko scaffold
}

