% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{mol2mw}
\alias{mol2mw}
\title{Calculate molecular weight}
\usage{
mol2mw(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
A list of molecular weights
}
\description{
Calculate molecular weight
}

