% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{mol2NumHBD}
\alias{mol2NumHBD}
\title{Calculate CalcNum Hidrogen bond acceptors}
\usage{
mol2NumHBD(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
NumRotatableBonds
}
\description{
Calculate CalcNum Hidrogen bond acceptors
}

