% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{mol2NumRotableBonds}
\alias{mol2NumRotableBonds}
\title{Calculate CalcNumRotatableBonds}
\usage{
mol2NumRotableBonds(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
NumRotatableBonds
}
\description{
Calculate CalcNumRotatableBonds
}

