% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{mol2smiles}
\alias{mol2smiles}
\title{Get Smiles of molecules}
\usage{
mol2smiles(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
A list of molecular weights
}
\description{
Get Smiles of molecules
}

