% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{molGetProps}
\alias{molGetProps}
\title{Map molecule properties to a list}
\usage{
molGetProps(m)
}
\arguments{
\item{m}{A molecule}
}
\value{
A list of properties
}
\description{
Map molecule properties to a list
}

