% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{molsGetProps}
\alias{molsGetProps}
\title{Get Properties of molecules}
\usage{
molsGetProps(mols)
}
\arguments{
\item{mols}{a list of molecules}
}
\value{
data.frame with properties
}
\description{
Get Properties of molecules
}

