% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{molsSetProps}
\alias{molsSetProps}
\title{Set Properties to Molecules}
\usage{
molsSetProps(ms, key = NULL, v = NULL, df = NULL)
}
\arguments{
\item{ms}{A molecule or list of molecules}

\item{key}{A key to be added or modified}

\item{v}{A value or vector}

\item{df}{A dataframe with columns to be included in the sdf}
}
\description{
Set Properties to Molecules
}

