% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_computeGasteigerCharges}
\alias{p_computeGasteigerCharges}
\title{Compute GasteigerCharges}
\usage{
p_computeGasteigerCharges(xp)
}
\arguments{
\item{xp}{a molecule}
}
\value{
the GasteigerCharges
}
\description{
Compute GasteigerCharges
}

