% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_SubstructMatch}
\alias{p_SubstructMatch}
\title{p_SubstructMatch}
\usage{
p_SubstructMatch(xp_mol, xp_query)
}
\arguments{
\item{xp_mol}{a molecule}

\item{xp_query}{a template molecule (from a smart)}
}
\value{
bool
}
\description{
p_SubstructMatch
}

