% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{read.sdf}
\alias{read.sdf}
\title{Read a SDF file and create molecules}
\usage{
read.sdf(file, sanitize = T, removeHs = T, strictParsing = T)
}
\arguments{
\item{file}{file name}

\item{sanitize}{- if true sanitize the molecule before returning it}

\item{removeHs}{- if true remove Hs from the molecule before returning it (triggers sanitization)}

\item{strictParsing}{- if not set, the parser is more lax about correctness of the contents.}
}
\value{
A list of molecules
}
\description{
Read a SDF file and create molecules
}

