% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{read.smi}
\alias{read.smi}
\title{Reads a smi file into a data.frame}
\usage{
read.smi(file, colnames = "")
}
\arguments{
\item{file}{File name}

\item{colnames}{Vector with colnames}
}
\value{
A data frame
}
\description{
Reads a smi file into a data.frame
}

