\name{RRDKit}
\alias{RRDKit}
\alias{RRDKit}
\docType{package}
\title{
A pragmatic interface to RDKit in R
}
\description{
This package provides a pragmatic interface to some of the RDKit functions in R. It is intended to work smoothly with R workflows.

RRDKit uses the RDKit installation in $RDBASE.

RRDKit provides some basic functions to visualize the molecules.
}
\details{
\tabular{ll}{
Package: \tab RRDKit\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2014-11-05\cr
License: \tab GPL-3\cr
}


}
\author{
Pau Carrio

Maintainer: Pau Carrio <paucarrio@gmail.com>
}
\references{
www.rdkit.org
}

\keyword{ rrdkit, molecule }
 
\examples{
%% ~~ simple examples of the most important functions ~~
}
