% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{showMols}
\alias{showMols}
\title{Show molecules as 2D in browser}
\usage{
showMols(mols, open = T, names = "", svg.size = 200, grid = FALSE)
}
\arguments{
\item{mols}{a list of molecules}

\item{open}{if TRUE show output in browser}

\item{names}{vector with mols ids}

\item{svg.size}{size of pictures}

\item{grid}{show molecules as grid}
}
\value{
Path to temporary generated html file
}
\description{
Show molecules as 2D in browser
}
\examples{
mols <- smiles2mol(c("CC(=O)NC1=CC=C(O)C=C1","CC(=O)OC1=CC=CC=C1C(O)=O"))
names <- c("Paracetamol", "Aspirin")

# showMols(mols, names=names)
}

