% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{showMolsDF}
\alias{showMolsDF}
\title{Show molecules as 2D with df in browser}
\usage{
showMolsDF(mols, df = data.frame(), open = T, svg.size = 200)
}
\arguments{
\item{mols}{a list of molecules}

\item{df}{a data frame}

\item{open}{if TRUE show output in browser}

\item{svg.size}{size of pictures}
}
\value{
Path to temporary generated html file
}
\description{
Show molecules as 2D with df in browser
}
\examples{

mols <- smiles2mol(c("CC(=O)NC1=CC=C(O)C=C1","CC(=O)OC1=CC=CC=C1C(O)=O"))
df <- data.frame(id=c(1,2),names=c("Paracetamol", "Aspirin"),mw=mol2mw(mols))
# showMolsDF(mols,df)
}

