% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{showMolsGrid}
\alias{showMolsGrid}
\title{Show molecules as 2D in browser}
\usage{
showMolsGrid(mols, group = 1, id = "", open = T, svg.size = 200)
}
\arguments{
\item{mols}{a list of molecules}

\item{group}{groups of molecules}

\item{id}{property id that should be used for naming molecules}

\item{open}{if TRUE show output in browser}

\item{svg.size}{size of pictures}
}
\value{
Path to temporary generated html file
}
\description{
Show molecules as 2D in browser
}
\examples{
mols <- smiles2mol(c("CC(=O)NC1=CC=C(O)C=C1","CC(=O)OC1=CC=CC=C1C(O)=O"))
names <- c("Paracetamol", "Aspirin")
showMolsGrid(mols)
}

