% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{smarts2mol}
\alias{smarts2mol}
\title{map a smarts to a molecule}
\usage{
smarts2mol(smarts, silent = FALSE)
}
\arguments{
\item{smarts}{a smarts string}

\item{silent}{if true ignore errors}
}
\value{
a molecule
}
\description{
map a smarts to a molecule
}

