% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SubstructMatch}
\alias{SubstructMatch}
\title{SubstructMatch}
\usage{
SubstructMatch(xp_mol, xp_query)
}
\arguments{
\item{xp_mol}{a molecule}

\item{xp_query}{a template molecule (from a smart)}
}
\value{
bool
}
\description{
SubstructMatch
}

