% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{tanimoto}
\alias{tanimoto}
\title{Calculate tanimoto distance of two bool vectors}
\usage{
tanimoto(l1, l2)
}
\arguments{
\item{l1}{a vector}

\item{l2}{a vector}
}
\value{
Tanimoto distance
}
\description{
Calculate tanimoto distance of two bool vectors
}

