% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{write.sdf}
\alias{write.sdf}
\title{Write a SDF file}
\usage{
write.sdf(file, mols, setForceV3000 = F)
}
\arguments{
\item{file}{File name}

\item{mols}{A list of molecules}

\item{setForceV3000}{force V3000}
}
\description{
Write a SDF file
}
\examples{

mols <- smiles2mol(c("CC(=O)NC1=CC=C(O)C=C1","CC(=O)OC1=CC=CC=C1C(O)=O"))
write.sdf("mols.sdf",mols)

compute2D(mols)
write.sdf("mols.sdf",mols)
}

