% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{computeGasteigerCharges}
\alias{computeGasteigerCharges}
\title{computeGasteigerCharges}
\usage{
computeGasteigerCharges(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
computeGasteigerCharges
}
\description{
computeGasteigerCharges
}

