% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{mol2murcko}
\alias{mol2murcko}
\title{Calculate murcko scaffold}
\usage{
mol2murcko(mols)
}
\arguments{
\item{mols}{A list of molecules}
}
\value{
murcko scaffolds
}
\description{
Calculate murcko scaffold
}

