% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{molsSetProps}
\alias{molsSetProps}
\title{Set Properties to Molecules}
\usage{
molsSetProps(ms, key, v)
}
\arguments{
\item{ms}{A molecule or list of molecules}

\item{key}{A key to be added or modified}

\item{v}{A value or vector}
}
\description{
Set Properties to Molecules
}

