% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{smiles2mol}
\alias{smiles2mol}
\title{Convert smiles to molecules}
\usage{
smiles2mol(smi, sanitize = TRUE)
}
\arguments{
\item{smi}{a smile string or vector of strings}

\item{sanitize}{toggles H removal and sanitization of the molecule}
}
\value{
List of molecules
}
\description{
Convert smiles to molecules
}
\examples{
mol <- smiles2mol("c1ccccc1")
mols <- smiles2mol(c("c1ccccc1","CC(=O)OC1=CC=CC=C1C(O)=O"))
}

