% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcppwrappers.R
\name{SubstructMatch}
\alias{SubstructMatch}
\title{p_SubstructMatch}
\usage{
SubstructMatch(mol, query)
}
\arguments{
\item{mol}{a molecule}

\item{query}{a template molecule (from a smart)}
}
\value{
bool
}
\description{
p_SubstructMatch
}

